Refactor .npy reader
